ingestion_service/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
prefect>=3.0.0
psycopg[binary]>=3.2.0
python-dateutil>=2.9.0
python-dotenv>=1.0.1
python-multipart>=0.0.20
